requirements.txt
streamlit
opencv-python
pillow
numpy
bcrypt